+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 
 # Next
+1. <font color=blue>增强</font> `url`以`http://`或`https://`开头的不再自动添加`prefix`